improvements (#8)
* remove numOfTiers

* fix initializer

* fuzz tests

* fix getUpdatedCurrentPeriod()

* add comments

* prevent frontrunning attack

* add nonce to prevent duplicated ids

* change periodDuration and currentPeriod to uint238

* fix errors

* comments + unsub bug

* uint128 fixes

* another bug

* formatting + change in how tiers are disabled/tiers are updated

* fix update()?

* fix claimable

---------

+++ src/LlamaSubsFactory.sol
@@ -15,8 +15,8 @@ contract LlamaSubsFactory {
     event DeployFlatRateERC20(
         address deployedContract,
         address indexed owner,
-        uint256 currentPeriod,
-        uint256 periodDuration,
+        uint128 currentPeriod,
+        uint128 periodDuration,
         LlamaSubsFlatRateERC20.TierInfo[] tiers
     );
     event DeployFlatRateERC20NonRefundable(
@@ -39,8 +39,8 @@ contract LlamaSubsFactory {
     }
 
     function deployFlatRateERC20(
-        uint256 _currentPeriod,
-        uint256 _periodDuration,
+        uint128 _currentPeriod,
+        uint128 _periodDuration,
         LlamaSubsFlatRateERC20.TierInfo[] memory tiers
     ) external returns (LlamaSubsFlatRateERC20 deployedContract) {
         deployedContract = LlamaSubsFlatRateERC20(
